# docs/openapi.yml
openapi: 3.1.0
info:
  title: Aegiscan API
  summary: Aegiscan API
  description: Aegiscan is the open source Tines / Splunk SOAR alternative. You can
    operate Aegiscan in headless mode by using the API to create, manage, and run
    workflows.
  termsOfService: https://docs.google.com/document/d/e/2PACX-1vQvDe3SoVAPoQc51MgfGCP71IqFYX_rMVEde8zC4qmBCec5f8PLKQRdxa6tsUABT8gWAR9J-EVs2CrQ/pub
  contact:
    name: Aegiscan Founders
    email: founders@digi-trans.org
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
  version: '1'
paths:
...
  /workspaces:
    get:
      tags:
      - workspaces
      summary: List Workspaces
      description: 'List workspaces.


        Access Level

        ------------

        - Basic: Can list workspaces where they are a member.

        - Admin: Can list all workspaces regardless of membership.'
      operationId: workspaces-list_workspaces
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/WorkspaceMetadataResponse'
                type: array
                title: Response Workspaces-List Workspaces
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
    post:
      tags:
      - workspaces
      summary: Create Workspace
      description: 'Create a new workspace.


        Access Level

        ------------

        - Admin: Can create a workspace for any user.'
      operationId: workspaces-create_workspace
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkspaceParams'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceMetadataResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
  /workspaces/search:
    get:
      tags:
      - workspaces
      summary: Search Workspaces
      description: Return Workflow as title, description, list of Action JSONs, adjacency
        list of Action IDs.
      operationId: workspaces-search_workspaces
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: name
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkspaceMetadataResponse'
                title: Response Workspaces-Search Workspaces
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /workspaces/{workspace_id}:
    get:
      tags:
      - workspaces
      summary: Get Workspace
      description: Return Workflow as title, description, list of Action JSONs, adjacency
        list of Action IDs.
      operationId: workspaces-get_workspace
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: workspace_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - workspaces
      summary: Update Workspace
      description: Update a workspace.
      operationId: workspaces-update_workspace
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: workspace_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWorkspaceParams'
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - workspaces
      summary: Delete Workspace
      description: Delete a workspace.
      operationId: workspaces-delete_workspace
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: workspace_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /workspaces/{workspace_id}/memberships:
    get:
      tags:
      - workspaces
      summary: List Workspace Memberships
      description: List memberships of a workspace.
      operationId: workspaces-list_workspace_memberships
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: workspace_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkspaceMembershipResponse'
                title: Response Workspaces-List Workspace Memberships
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - workspaces
      summary: Create Workspace Membership
      description: Create a workspace membership for a user.
      operationId: workspaces-create_workspace_membership
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: workspace_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkspaceMembershipParams'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /workspaces/{workspace_id}/memberships/{user_id}:
    get:
      tags:
      - workspaces
      summary: Get Workspace Membership
      description: Get a workspace membership for a user.
      operationId: workspaces-get_workspace_membership
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: workspace_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid4
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceMembershipResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - workspaces
      summary: Delete Workspace Membership
      description: Delete a workspace membership.
      operationId: workspaces-delete_workspace_membership
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: workspace_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid4
          title: User Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CreateWorkspaceMembershipParams:
      properties:
        user_id:
          type: string
          format: uuid4
          title: User Id
      type: object
      required:
      - user_id
      title: CreateWorkspaceMembershipParams
    CreateWorkspaceParams:
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 1
          title: Name
        settings:
          anyOf:
          - additionalProperties:
              type: string
            type: object
          - type: 'null'
          title: Settings
        owner_id:
          type: string
          format: uuid
          title: Owner Id
          default: 00000000-0000-0000-0000-000000000000
      type: object
      required:
      - name
      title: CreateWorkspaceParams
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    UpdateWorkspaceParams:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 100
            minLength: 1
          - type: 'null'
          title: Name
        settings:
          anyOf:
          - additionalProperties:
              type: string
            type: object
          - type: 'null'
          title: Settings
      type: object
      title: UpdateWorkspaceParams
...
    UserRole:
      type: string
      enum:
      - basic
      - admin
      title: UserRole
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    WorkspaceMember:
      properties:
        user_id:
          type: string
          format: uuid4
          title: User Id
        first_name:
          anyOf:
          - type: string
          - type: 'null'
          title: First Name
        last_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Name
        email:
          type: string
          format: email
          title: Email
        role:
          $ref: '#/components/schemas/UserRole'
      type: object
      required:
      - user_id
      - first_name
      - last_name
      - email
      - role
      title: WorkspaceMember
    WorkspaceMembershipResponse:
      properties:
        user_id:
          type: string
          format: uuid4
          title: User Id
        workspace_id:
          type: string
          format: uuid4
          title: Workspace Id
      type: object
      required:
      - user_id
      - workspace_id
      title: WorkspaceMembershipResponse
    WorkspaceMetadataResponse:
      properties:
        id:
          type: string
          format: uuid4
          title: Id
        name:
          type: string
          title: Name
        n_members:
          type: integer
          title: N Members
      type: object
      required:
      - id
      - name
      - n_members
      title: WorkspaceMetadataResponse
    WorkspaceResponse:
      properties:
        id:
          type: string
          format: uuid4
          title: Id
        name:
          type: string
          title: Name
        settings:
          anyOf:
          - additionalProperties:
              type: string
            type: object
          - type: 'null'
          title: Settings
        owner_id:
          type: string
          format: uuid
          title: Owner Id
        n_members:
          type: integer
          title: N Members
        members:
          items:
            $ref: '#/components/schemas/WorkspaceMember'
          type: array
          title: Members
      type: object
      required:
      - id
      - name
      - owner_id
      - n_members
      - members
      title: WorkspaceResponse
...
  securitySchemes:
    APIKeyCookie:
      type: apiKey
      in: cookie
      name: fastapiusersauth
    APIKeyHeader:
      type: apiKey
      in: header
      name: x-aegiscan-service-key
...